requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
coverage==7.3.1
decorator==5.1.1
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
factory-boy==3.3.0
Faker==19.6.1
importlib-metadata==6.8.0
ipython==8.15.0
jedi==0.19.0
Markdown==3.4.4
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
typing_extensions==4.7.1
wcwidth==0.2.6
zipp==3.16.2
